requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.3.1
matplotlib>=3.4.3
seaborn>=0.11.2
scikit-learn>=0.24.2
tqdm>=4.62.3
pandas>=1.3.0 